rounding scheme overhauled to avoid having equal start and end weight in extreme small cases

+++ src/main/java/com/cropster/roastingsimulation/common/random/RandomGenerationService.java
@@ -14,13 +14,13 @@ public interface RandomGenerationService {
 
     int getRandomInitialGreenCoffeeAmount();
 
-    int getRandomRoastingStartWeight(int machineCapacity);
+    double getRandomRoastingStartWeight(int machineCapacity);
 
     int getRandomDuraion();
 
     Date getRandomStartTime();
 
-    int getRandomWeightLoss(int startWeight);
+    double getRandomWeightLoss(double startWeight);
 
     String getRandomProductName();
 }
